pyproject.toml
[tool.poetry]
name = "zoomcamp24"
version = "0.1.0"
description = ""
authors = ["Ezequiel Lopez <ezequiel20139@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.1.4"
pgcli = "^4.0.1"
sqlalchemy = "^2.0.25"
wget = "^3.2"
fastparquet = "^2023.10.1"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
